% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsam-deprecated.R
\name{fitSSM}
\alias{diagSSM}
\alias{fitSSM}
\alias{plotSSM}
\title{Deprecated functions.}
\usage{
fitSSM(...)

diagSSM(...)

plotSSM(...)
}
\arguments{
\item{...}{ignored}
}
\description{
\code{fitSSM}, \code{diagSSM}, and \code{plotSSM}, have been deprecated. Instead use
\code{fit_ssm}, \code{diag_ssm} and \code{map_ssm}.
}

