% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag_ssm.R
\name{diag_ssm}
\alias{diag_ssm}
\title{Plot standard McMC convergence diagnostics to help determine lack of model convergence.}
\usage{
diag_ssm(fit)
}
\arguments{
\item{fit}{an output object from \code{fitSSM}}
}
\value{
The traceplot shows the time series for both McMC chains;
the density plot shows the density estimate for each parameter;
the autocorrelation plots show the within-chain sample autocorrelation for each parameter;
the G-B-R shrink factor plot shows the evolution of Gelman and Rubin's shrink factor for 
increasing number of iterations. See the \code{coda} package for further details.
}
\description{
Takes a fitted \code{fitSSM} object and for each of the parameters plots:
}
\examples{
\dontrun{
data(ellie)
fit <- fitSSM(ellie, model = "DCRWS", tstep = 1, adapt = 2000, samples = 1000, 
              thin = 2, span = 0.1)
diag_ssm(fit)

# increase burnin, posterior sample numbers, and thinning factor
fit2 <- fitSSM(ellie, model = "DCRWS", tstep = 1, adapt = 5000, samples = 5000, 
              thin = 5, span = 0.1)
diag_ssm(fit2)
}             
}
\references{
Brooks SP, Gelman A (1998) General methods for monitoring convergence of iterative 
simulations. Journal of Computational and Graphical Statistics 7:434-455
}

