% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsam-package.R
\docType{data}
\name{ellie2}
\alias{ellie2}
\title{Elephant seal Argos satellite data (2 individuals)}
\format{
.RData
}
\description{
Example elephant seal Argos tracking data. Data were sourced from 
the Integrated Marine Observing System (IMOS) - IMOS is supported by the 
Australian Government through the National Collaborative Research Infrastructure 
Strategy and the Super Science Initiative.
}
\keyword{data}
