% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hssm.R
\name{hssm}
\alias{hssm}
\title{hssm}
\usage{
hssm(loc.list, model = "hDCRWS", adapt, samples, thin, chains, ...)
}
\arguments{
\item{loc.list}{loc.list}

\item{model}{model}

\item{adapt}{adapt}

\item{samples}{samples}

\item{thin}{thin}

\item{chains}{chains}

\item{...}{dots}
}
\description{
hssm
}

