% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.r
\name{simulate}
\alias{simulate}
\title{Simulate from the DCRW model with Argos location errors}
\usage{
simulate(
  Nt = 100,
  gamma = 0.8,
  Sigma = matrix(c(0.01, 0, 0, 0.01), 2, 2),
  amf = tpar()
)
}
\arguments{
\item{Nt}{number of time steps to simulate}

\item{gamma}{move persistence parameter}

\item{Sigma}{variance-covariance matrix for movement process}

\item{amf}{Argos error data, defined by default via the \code{tpar} function which
uses the t-distribution scale and df estimates from Jonsen et al (2005)}
}
\value{
a data_frame of true locations and locations with Argos error
}
\description{
For testing bsam models
}
