% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpar.R
\name{tpar}
\alias{tpar}
\title{ARGOS Error Fixed Parameters}
\usage{
tpar()
}
\value{
A dataframe with columns
\item{\code{lc}}{ARGOS location class}
\item{\code{itau2.lon}}{precision parameters for longitude}
\item{\code{itau2.lat}}{precision parameters for latitude}
\item{\code{nu.lon}}{df parameters for longitude}
\item{\code{nu.lat}}{df parameters for latitude}
}
\description{
ARGOS Error Fixed Parameters for Location Classes
}
\details{
This is an internal function used by \code{dat4jags} to specify measurement
  error parameters.

These are the fixed parameters (t-distribution scale & df) for ARGOS
error classes, from Jonsen et al (2005) Ecology 86:2874-2880
}

