% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tplot.R
\name{tplot}
\alias{tplot}
\title{Plot estimated track, behavioural states and observations on a map.}
\usage{
tplot(fit, onemap = TRUE)
}
\arguments{
\item{fit}{an output object from \code{fitSSM}}

\item{onemap}{logical (default is FALSE) indicating if individual tracks
are combined on a single map (TRUE) or on separate maps (FALSE)}
}
\description{
Estimated location states are shaded by estimated behavioural state 
(for DCRWS, hDCRWS output). Observations are added as '+' symbols. 
For multi-individual output objects there is an option to include all
on a single map or to plot each individually.
}

