% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat4jags.R
\name{dat4jags}
\alias{dat4jags}
\title{Prepares and writes input data for DCRW/DCRWS/hDCRWS models}
\usage{
dat4jags(indata, tstep = 1, ...)
}
\arguments{
\item{indata}{An R data.frame of Argos tracking data. See fitSSM (indata)
for details on structure.}

\item{tstep}{The time step to be assumed for the state-space model,
specified in days.}

\item{\dots}{Other arguments may be passed.}
}
\value{
Returns a list to be used by ssm/hssm.
}
\description{
Takes an R data.frame of Argos tracking data restructures it for the models.
Intended for internal use, called by fitSSM.
}
\seealso{
Function to be called by \code{\link{fitSSM}}.
}

