% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hssm.R
\name{hssm}
\alias{hssm}
\title{hssm}
\usage{
hssm(loc.list, model = "hDCRWS", adapt, samples, thin, chains, ...)
}
\arguments{
\item{loc.list}{}

\item{model}{}

\item{adapt}{}

\item{samples}{}

\item{thin}{}

\item{chains}{}

\item{...}{}
}
\description{
hssm
}

