% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simTrack.R
\name{simTrack}
\alias{simTrack}
\title{simTrack}
\usage{
simTrack(T = 100, theta = c(0, pi), gamma = c(0.9, 0.2), alpha = c(0.2,
  0.8), vcov = c(0.01, 0, 0, 0.02))
}
\arguments{
\item{theta}{}

\item{gamma}{}

\item{alpha}{}

\item{vcov}{}
}
\description{
simTrack
}

