% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm.R
\name{ssm}
\alias{ssm}
\title{Fits state-space model to Argos data}
\usage{
ssm(loc.list, model = "DCRW", adapt, samples, thin, chains, ...)
}
\arguments{
\item{loc.list}{Data from dat4jags to be passed to JAGS}

\item{model}{The state-space model to be fit. Currently, DCRW and DCRWS are}

\item{adapt}{adapt}

\item{samples}{samples}

\item{thin}{thin}

\item{chains}{chains}

\item{...}{the only model options.}
}
\value{
Returns a list of McMC samples from marginal posteriors and a
summary data.frame of mean and median position estimates.
}
\description{
Takes output from dat4jags, sets up initial values, calls JAGS, and
aggregates results. Intended for internal use, called by fitSSM.
}
\seealso{
Function to be called by \code{\link{fitSSM}}.
}

